adds a subgraph template that maps erc20 frax contracts the token mapping controls the underlying fractionalized nft's supply
