+++ subgraph/schema.graphql
@@ -19,7 +19,7 @@ type Fractionalized @entity {
   #initialAmount: BigInt! # initial amount of tokens
   totalIssued: BigInt! #the highest amount of fractions issued, this can be raised by the owner
   circulatingSupply: BigInt! #the amount of unburnt fractions
-  erc20address: Bytes #address of the underlying erc20 token
+  erc20address: Bytes! #address of the underlying erc20 token
   symbol: String
   paymentToken: Bytes #address ERC20 Token
   paidPrice: BigInt #the price paid for the original ipnft
